Remove brain reply commands; add feed edit-reply/delete-reply; rename feed reply → post-reply
Brain command group now only manages the user's own posts (post-update,
edit-update, delete-update). All reply interactions move under feed:
post-reply, edit-reply, delete-reply, plus the existing list/get. The
post-reply name aligns with brain's post-update convention. Reply
edit/delete still hit /brain-updates/replies/:id since the feed
controller doesn't expose them yet.
