+++ .claude-plugin/plugin.json
@@ -1,7 +1,7 @@
 {
   "name": "gobi",
   "description": "Manage the Gobi collaborative knowledge platform from the command line",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "author": {
     "name": "gobi-ai"
   },
